feature - 14 - refactor quito "test"
