+++ src/main/java/com/cap/pricecontrolcenter/PriceControlCenterApplication.java
@@ -1,0 +1,13 @@
+package com.cap.pricecontrolcenter;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PriceControlCenterApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PriceControlCenterApplication.class, args);
+	}
+
+}
