+++ src/main/java/com/cap/pricecontrolcenter/domain/model/PriceModel.java
@@ -1,0 +1,21 @@
+package com.cap.pricecontrolcenter.domain.model;
+
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+@Getter
+@Builder
+public class PriceModel {
+    private Integer brandId;
+    private LocalDateTime startDate;
+    private LocalDateTime endDate;
+    private Integer priceList;
+    private Integer productId;
+    private Integer priority;
+    private BigDecimal price;
+    private String currency;
+}
