+++ src/main/java/com/cap/pricecontrolcenter/infraestructure/exception/GlobalExceptionHandler.java
@@ -1,0 +1,37 @@
+package com.cap.pricecontrolcenter.infraestructure.exception;
+
+import com.cap.pricecontrolcenter.infraestructure.exception.custom.PriceCreationException;
+import java.util.Map;
+import java.util.StringJoiner;
+import java.util.stream.Collectors;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.MissingServletRequestParameterException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<Map<String, String>> handleValidationException(MethodArgumentNotValidException ex) {
+        Map<String, String> errors = ex.getBindingResult().getFieldErrors()
+                .stream()
+                .collect(Collectors.toMap(FieldError::getField, FieldError::getDefaultMessage));
+        return ResponseEntity.badRequest().body(errors);
+    }
+
+    @ExceptionHandler(MissingServletRequestParameterException.class)
+    public ResponseEntity<String> handleMissingServletRequestParameterException(MissingServletRequestParameterException ex) {
+        StringJoiner joiner = new StringJoiner(", ", "Missing request parameters: ", "");
+        joiner.add(ex.getParameterName());
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(joiner.toString());
+    }
+
+    @ExceptionHandler(PriceCreationException.class)
+    public ResponseEntity<String> handlePriceCreationException(PriceCreationException ex) {
+        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An error occurred while creating the price.");
+    }
+}
