+++ src/main/java/com/cap/pricecontrolcenter/infraestructure/exception/custom/PriceCreationException.java
@@ -1,0 +1,7 @@
+package com.cap.pricecontrolcenter.infraestructure.exception.custom;
+
+public class PriceCreationException extends RuntimeException {
+    public PriceCreationException() {
+        super("Error while creating the price");
+    }
+}
