--- src/main/java/com/test/cap/pricecontrolcenter/PriceControlCenterApplication.java
@@ -1,13 +1,0 @@
-package com.test.cap.pricecontrolcenter;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PriceControlCenterApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PriceControlCenterApplication.class, args);
-	}
-
-}
