--- src/main/java/com/test/cap/pricecontrolcenter/common/UseCase.java
@@ -1,27 +1,0 @@
-package com.test.cap.pricecontrolcenter.common;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import org.springframework.core.annotation.AliasFor;
-import org.springframework.stereotype.Component;
-
-
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-@Component
-public @interface UseCase {
-
-    /**
-     * The value may indicate a suggestion for a logical component name,
-     * to be turned into a Spring bean in case of an autodetected component.
-     * @return the suggested component name, if any (or empty String otherwise)
-     */
-    @AliasFor(annotation = Component.class)
-    String value() default "";
-
-}
