--- src/main/java/com/test/cap/pricecontrolcenter/infraestructure/exception/custom/PriceCreationException.java
@@ -1,7 +1,0 @@
-package com.test.cap.pricecontrolcenter.infraestructure.exception.custom;
-
-public class PriceCreationException extends RuntimeException {
-    public PriceCreationException() {
-        super("Error while creating the price");
-    }
-}
